==> 02_prepare_models.Rmd <==
---
title: "prepare models"
author: "SPT"
date: "2022_05_06"
output: github_document
---

### Step 1: Load packages and data

```{r, echo = `FALSE`}
# Load packages
library(tidyverse)
library(caret)
library(dplyr)
library(readxl)
library(data.table)

# Reset global environment
rm(list = ls())

# Change working directory
setwd("/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp/data")

# Load data
df.historical <- fread("processed/df_historical.csv")
df.future     <- fread("processed/df_future.csv")

# Air variables
var.air.raw     <- c("max_air", "mean_air", "min_air")
var.air.z       <- c("max_air_z", "mean_air_z", "min_air_z")
var.air.other   <- c("air_quad_temp", "precip", "ice_duration")
var.air         <- c(var.air.raw,
                     var.air.z,
                     var.air.other)

# Water variables
var.water.raw   <- c("max_bot", "max_surf", "mean_bot", "mean_surf")
var.water.z     <- c("max_surf_z", "mean_surf_z", "max_bot_z", "mean_bot_z")
var.water.other <- c("peak_temp", "layer_diff", "water_quad_temp",
                     "variance_after_ice_30", "variance_after_ice_60", "schmidt",
                     "cumulative_above_0", "cumulative_above_5", "cumulative_above_10") #"precip", "ice_duration")
var.water.s     <- c("max_surf", "mean_surf", "max_surf_z", "mean_surf_z")
var.water.b     <- c("max_bot", "mean_bot", "max_bot_z", "mean_bot_z")
var.water       <- c(var.water.raw,
                     var.water.z,
                     var.water.other)

# Productivity and lake size variables
var.lake        <- c("secchi_m", "size_km2")

# All environmental variables
var.all.env     <- c(var.air, var.water, var.lake)

# Fish variables
var.fish   <- c("centrarchidae", "percidae", "esocidae", "ictaluridae",
                "cyprinidae", "salmonidae", "catostomidae", "sciaenidae",
                "clupeidae", "osmeridae", "lepisosteidae", "acipenseridae",
                "amiidae", "cottoidae", "gasterosteidae", "gobiidae",
                "fish_families", "fish_species")

# Cause variables
var.cause <- c("anthropogenic", "infectious", "summerkill", "winterkill", "unknown")          

# Fish thermal variables
var.temp  <- c("cold_temp", "cool_temp", "warm_temp", "unknown_temp")

# All variables
var.all.fish <-c(var.fish, var.temp, var.cause)

######################################
# Reduce historical dataframe for models
df.historical <- df.historical %>%
    mutate(
      year         = as.character(year),
      month        = as.character(month),
      season       = as.factor(season),
      summerkill   = as.numeric(summerkill),
      lat          = as.numeric(lat),
      long         = as.numeric(long),
      state        = as.factor(state),
      population   = as.numeric(population),
      ice_duration = as.numeric(ice_duration),
      secchi_m     = as.numeric(secchi_m),
      size_km2     = as.numeric(size_km2)) %>%
  dplyr::select(., c(year, month, season, summerkill, lat, long, state, population,
                     all_of(var.all.env), all_of(var.all.fish))) %>%
  mutate(summerkill = ifelse(is.na(summerkill) == TRUE, 0, summerkill)) %>%
  mutate(summerkill = as.factor(as.character(summerkill))) %>%
  mutate(summerkill = fct_recode(summerkill,
                                 "pos" = "1",
                                 "neg" = "0"))

######################################
# Reduce future dataframe for models
df.future <- df.future %>%
    mutate(
      year         = as.character(year),
      month        = as.character(month),
      season       = as.factor(season),
      lat          = as.numeric(lat),
      long         = as.numeric(long),
      state        = as.factor(state),
      population   = as.numeric(population),
      ice_duration = as.numeric(ice_duration)) %>%
  dplyr::select(., c(year, month, season, lat, long, state, population,
                     all_of(var.all.env)))
```

### Step 3: Historical PCAs

```{r echo = `FALSE`}
#######################################
# Air (raw and z-score)

# PCA
air.both.pca <- df.historical %>%
  dplyr::select(all_of(c(var.air.raw, var.air.z))) %>%
  prcomp(center = TRUE, scale = TRUE)

# Proportion of variance explained by each rotation
air.both.pca$rotation

# Calculate proportion of variance explained
round((air.both.pca$sdev / sum(air.both.pca$sdev)), 2)

# Separate PC1:2
air_pca <- df.historical %>%
  dplyr::select(max_air,   mean_air,   min_air,
                max_air_z, mean_air_z, min_air_z) %>%
  as.matrix() %*% air.both.pca$rotation[, 1:2]

# Add PCA results into monthly water temperature data
df.historical <- df.historical %>%
  add_column(air_pc1 = air_pca[,1],
             air_pc2 = air_pca[,2])

######################################
# Water (raw and z-score)

# PCA
water.both.pca <- df.historical %>%
  dplyr::select(all_of(c(var.water.raw, var.water.z))) %>%
  prcomp(center = TRUE, scale = TRUE)

# Proportion of variance explained by each rotation
water.both.pca$rotation

# Calculate proportion of variance explained
round((water.both.pca$sdev / sum(water.both.pca$sdev)), 2)

# Separate PC1:3
water_pca <- df.historical %>%
  dplyr::select(max_surf,   mean_surf,   max_bot,   mean_bot,
                max_surf_z, mean_surf_z, max_bot_z, mean_bot_z) %>%
  as.matrix() %*% water.both.pca$rotation[, 1:3]

# Add PCA results into monthly water temperature data
df.historical <- df.historical %>%
  add_column(water_pc1 = water_pca[,1],
             water_pc2 = water_pca[,2],
             water_pc3 = water_pca[,3])
```

### Step 4: Future (mid) PCAs

```{r echo = `FALSE`}
# Change summerkill back to numeric for PCAs
df.future.mid <- df.future %>%
  filter(year < 2075)

#######################################
# Air (raw and z-score)

# PCA
air.both.pca <- df.future.mid %>%
  dplyr::select(all_of(c(var.air.raw, var.air.z))) %>%
  prcomp(center = TRUE, scale = TRUE)

# Proportion of variance explained by each rotation
air.both.pca$rotation

# Calculate proportion of variance explained
round((air.both.pca$sdev / sum(air.both.pca$sdev)), 2)

# Separate PC1:2
air_pca <- df.future.mid %>%
  dplyr::select(max_air,   mean_air,   min_air,
                max_air_z, mean_air_z, min_air_z) %>%
  as.matrix() %*% air.both.pca$rotation[, 1:2]

# Add PCA results into monthly water temperature data
df.future.mid <- df.future.mid %>%
  add_column(air_pc1 = air_pca[,1],
             air_pc2 = air_pca[,2])

######################################
# Water (raw and z-score)

# PCA
water.both.pca <- df.future.mid %>%
  dplyr::select(all_of(c(var.water.raw, var.water.z))) %>%
  prcomp(center = TRUE, scale = TRUE)

# Proportion of variance explained by each rotation
water.both.pca$rotation

# Calculate proportion of variance explained
round((water.both.pca$sdev / sum(water.both.pca$sdev)), 2)

# Separate PC1:3
water_pca <- df.future.mid %>%
  dplyr::select(max_surf,   mean_surf,   max_bot,   mean_bot,
                max_surf_z, mean_surf_z, max_bot_z, mean_bot_z) %>%
  as.matrix() %*% water.both.pca$rotation[, 1:3]

# Add PCA results into monthly water temperature data
df.future.mid <- df.future.mid %>%
  add_column(water_pc1 = water_pca[,1],
             water_pc2 = water_pca[,2],
             water_pc3 = water_pca[,3])
```

### Step 5: Future (late) PCAs

```{r echo = `FALSE`}
# Change summerkill back to numeric for PCAs
df.future.late <- df.future %>%
  filter(year > 2075)

#######################################
# Air (raw and z-score)

# PCA
air.both.pca <- df.future.late %>%
  dplyr::select(all_of(c(var.air.raw, var.air.z))) %>%
  prcomp(center = TRUE, scale = TRUE)

# Proportion of variance explained by each rotation
air.both.pca$rotation

# Calculate proportion of variance explained
round((air.both.pca$sdev / sum(air.both.pca$sdev)), 2)

# Separate PC1:2
air_pca <- df.future.late %>%
  dplyr::select(max_air,   mean_air,   min_air,
                max_air_z, mean_air_z, min_air_z) %>%
  as.matrix() %*% air.both.pca$rotation[, 1:2]

# Add PCA results into monthly water temperature data
df.future.late <- df.future.late %>%
  add_column(air_pc1 = air_pca[,1],
             air_pc2 = air_pca[,2])

######################################
# Water (raw and z-score)

# PCA
water.both.pca <- df.future.late %>%
  dplyr::select(all_of(c(var.water.raw, var.water.z))) %>%
  prcomp(center = TRUE, scale = TRUE)

# Proportion of variance explained by each rotation
water.both.pca$rotation

# Calculate proportion of variance explained
round((water.both.pca$sdev / sum(water.both.pca$sdev)), 2)

# Separate PC1:3
water_pca <- df.future.late %>%
  dplyr::select(max_surf,   mean_surf,   max_bot,   mean_bot,
                max_surf_z, mean_surf_z, max_bot_z, mean_bot_z) %>%
  as.matrix() %*% water.both.pca$rotation[, 1:3]

# Add PCA results into monthly water temperature data
df.future.late <- df.future.late %>%
  add_column(water_pc1 = water_pca[,1],
             water_pc2 = water_pca[,2],
             water_pc3 = water_pca[,3])

# Join updated mid- and late-future datasets
df.future <- rbind(df.future.mid, df.future.late)
```

### Step 5: Training and testing sets

Formats and partitions the data into training and testing sets, which are scaled to allow for coefficient comparison.

```{r echo = `FALSE`}
# Reduce historical dataframe for models
df.historical <- df.historical %>%
    mutate(
      year         = as.character(year),
      month        = as.character(month),
      season       = as.factor(season),
      summerkill   = as.factor(summerkill),
      lat          = as.numeric(lat),
      long         = as.numeric(long),
      state        = as.factor(state),
      population   = as.numeric(population),
      ice_duration = as.numeric(ice_duration)) %>%
  dplyr::select(., c(year, month, season, summerkill,
                     lat, long, state, population,
                     all_of(var.all.env),
                     water_pc1, water_pc2, water_pc3,
                     air_pc1, air_pc2))
```

```{r echo = `FALSE`}
# Set seed
set.seed(406)

# Creates series of training partitions
in.training <- createDataPartition(df.historical$summerkill, p = .75, list = FALSE)

# Partition training data
training <- df.historical  %>%
  slice(in.training)

# Partition testing data
testing <- df.historical %>%
  slice(-in.training)

# Scale data
pre.proc.values <- preProcess(training, method = c("center", "scale"))

# Create training sets
training <- predict(pre.proc.values, training) %>%
  mutate(year = as.integer(year))

# Create testing sets
testing <- predict(pre.proc.values, testing) %>%
  mutate(summerkill = factor(testing$summerkill, 
                             levels = c("pos", "neg"),
                             ordered = TRUE),
         year = as.integer(year))

# Export data
write.csv(testing,       "models/testing.csv")
write.csv(training,      "models/training.csv")
write.csv(df.historical, "processed/df_historical2.csv")
write.csv(df.future,     "processed/df_future2.csv")
```

######################################
Proceed to `03_fit_models.Rmd`
######################################
