==> 04_assess_models.Rmd <==
---
title: "assess models"
author: "SPT"
date: "2022_05_06"
output: github_document
---

Candidate models are compared in terms of their coefficients to check for sensitivity of the coefficients to modeling choices, and their out-of-sample predictive performance, assessed using logloss and ROC-AUC.

### Step 1: Load packages and data

```{r, setup, include = `FALSE`, echo = `FALSE`}
knitr::opts_knit$set(root.dir = "/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp")
```

```{r, echo = `FALSE`}
# Load packages
library(tidyverse)
library(caret)
library(glmnet)
library(Matrix)
library(e1071)
library(broom)
library(yardstick)
library(purrr)
library(data.table)

# Reset global environment
rm(list = ls())

# Set working directory
setwd("/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp/data/models")

# Load data
training    <- fread("training.csv")
testing     <- fread("testing.csv")
```

### Step 2: Verify formatting

```{r, echo = `FALSE`} 
# Reformat training data
training <- training %>%
  mutate(year         = as.character(year),
         month        = as.character(month),
         season       = as.factor(season),
         summerkill   = as.factor(summerkill),
         lat          = as.numeric(lat),
         long         = as.numeric(long),
         state        = as.character(state),
         population   = as.numeric(population),
         mean_bot_z   = as.numeric(mean_bot_z),
         mean_surf_z  = as.numeric(mean_surf_z),
         mean_air_z   = as.numeric(mean_air_z),
         ice_duration = as.numeric(ice_duration))
  
# Reformat testing data
testing <- testing %>%
  mutate(year         = as.character(year),
         month        = as.character(month),
         season       = as.factor(season),
         summerkill   = as.factor(summerkill),
         lat          = as.numeric(lat),
         long         = as.numeric(long),
         state        = as.character(state),
         population   = as.numeric(population),
         mean_bot_z   = as.numeric(mean_bot_z),
         mean_surf_z  = as.numeric(mean_surf_z),
         mean_air_z   = as.numeric(mean_air_z),
         ice_duration = as.numeric(ice_duration))
```

### Step 2: Load models

```{r, setup, include = `FALSE`, echo = `FALSE`}
knitr::opts_knit$set(root.dir = "/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp/data/models/assessment")
```

```{r, echo = `FALSE`}
# Set working directory
setwd("/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp/data/models/assessment")

# Save file names
model_names <- list.files()

# Load full models
all_model_names <- model_names %>%
  as.data.frame(.) %>%
  dplyr::rename(model = 1)

all_models <- as.vector(all_model_names$model) %>%
  purrr::map(., ~read_rds(.))

# Extract model names and objects
all_models <- tibble(name = str_sub(str_extract(all_model_names$model, "^.*\\."),
                                    1, -2), model = all_models)
```

### Step 3: Extract fitted values

Extract fitted values and append as new column.

```{r, echo = `TRUE`}
# Function to compare fitted values on testing data
predict_route <- function(model, newdata) {
  if("train" %in% class(model)) {
    stats::predict(model, newdata, type = "prob")$pos
  } else {stats::predict(model, newdata, type = "response")}
}

# Add comparison results
all_models <- all_models %>%
  mutate(fit_test = purrr::map(model, predict_route, newdata = testing))

# Calculate logloss and auc for full models
all.auc <- rep(NA, nrow(all_models))
all.logloss <- all.auc
for (i in 1:nrow(all_models)) {
  a <- tibble(fit_test = all_models$fit_test[[i]],
              truth    = testing$summerkill)
  all.auc[i] <- roc_auc(a, truth, fit_test)$.estimate
  all.logloss[i] <- mn_log_loss(a, truth, fit_test)$.estimate
}

# Add logloss and auc for full models
all_models <- all_models %>%
  add_column(all.auc) %>%
  add_column(all.logloss) %>%
  dplyr::rename(auc     = all.auc,
                logloss = all.logloss) %>%
  arrange(logloss, desc(auc))

# Print logloss and auc in decreasing order
print(tibble::as_tibble(all_models[c("name", "logloss")]), n = 36)
# Top candidate models based on auc and logloss
# 1 ridge_a2_ds             2.39
# 2 ridge_w1_ds             2.44
# 3 ridge_w2_ds             2.45
# 4 lasso_w3_ds             2.59
```

### Step 4: Extract coefficients

Extract coefficients from each model.

```{r, echo = `FALSE`}
# Function for extracting coefficients from different model types
extract_coef <- function(model, name) {
  if ("train" %in% class(model)) {
    m <- model$finalModel %>%
      coef(model$bestTune$lambda)
    out <- tibble(term = row.names(m),
                  estimate = m[, 1]) %>%
      filter(term != "(Intercept)") %>%
      dplyr::select(term, estimate)
  }
  if ("glmerMod" %in% class(model)) {
    out <- model %>%
      tidy() %>%
      filter(group == "fixed") %>%
      filter(term != "(Intercept)") %>%
      filter(p.value <= .3) %>%
      dplyr::select(term, estimate)
  }
  if ("glm" %in% class(model)) {
    out <- model %>%
      tidy() %>%
      filter(term != "(Intercept)") %>%
      dplyr::select(term, estimate)
  }
  out
}

# Add coefficients
all_models <- all_models %>%
  mutate(coef = map2(model, name, extract_coef))

# Subset models
a1_models  <- all_models[grep("a1", x = all_models$name), ]
a2_models  <- all_models[grep("a2", x = all_models$name), ]
w1_models  <- all_models[grep("w1", x = all_models$name), ]
w2_models  <- all_models[grep("w2", x = all_models$name), ]

## Subset best models
best_models <- all_models %>%
  subset(., name == "ridge_a1_ds" | name == "ridge_a2_ds" | name == "ridge_w1_ds" | name == "ridge_w2_ds")
```

### Step 5: Plot coefficients (Fig. S1)

Plot coefficients of lesser models

```{r, setup, include = `FALSE`, echo = `FALSE`}
knitr::opts_knit$set(root.dir = "/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp/figures")
```

```{r, echo = `FALSE`}
# Set working directory
setwd("/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp/figures")

##################

# Describe PDF output
pdf(file = "fig_S1.pdf", width  = 8, height = 8)

# Plot coefficients of best fit models
best_models %>%
  unnest(coef) %>%
  ggplot(aes(x = fct_reorder(term, estimate), 
             y = estimate, 
             fill = estimate > 0)) +
  geom_col(show.legend = FALSE) +
  scale_x_discrete(position = "top") +
  facet_wrap(~ name, ncol = 2) +
  coord_flip() +
  labs(x = NULL) +
  theme_bw()

# Close plot
dev.off()
```

######################################
Proceed to `05_fit_full_models.Rmd`
######################################
