==> 06_statistics.Rmd <==
---
title: "statistics"
author: "SPT"
date: "2022_05_06"
output: github_document
---

### Step 1: Load packages and data

```{r, setup, include = `FALSE`, echo = `FALSE`}
knitr::opts_knit$set(root.dir = "/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp")
```

```{r, echo = `FALSE`}
# Load packages
library(ggplot2)
library(ggthemes)
library(ggmap)
library(gridExtra)
library(scales)
library(car)
library(DescTools)
library(spdep)
library(readr)
library(ggpubr)
library(rstatix)
library(reshape2)
library(maddog)
library(dplyr)
library(fmsb)
library(rstatix)
library(caret)
library(data.table)
library(tidyr)
#library(RVAideMemoire)

# Reset global environment
rm(list = ls())

# Change working directory
setwd("/Users/simontye/Research/Projects/MME_Temp/2020_MME_Temp")

# Load data
df.historical    <- fread("data/processed/df_historical.csv")
df.historical2   <- fread("data/processed/df_historical2.csv")
df.future        <- fread("data/processed/df_future2.csv")
df.snow          <- fread("data/processed/df_snow.csv")
thermal          <- fread("data/raw/thermal/fish_thermal.csv")
time.events      <- fread("data/raw/time/time_events.csv")
time.summerkills <- fread("data/raw/time/time_summerkills.csv")

# Load final models
ridge_a2 <- readRDS("data/models/full/full_ridge_a2_ds.rds")
ridge_w1 <- readRDS("data/models/full/full_ridge_w1_ds.rds")

# Evaluate models
eval_results <- function(true, predicted, df) {
  # Calculate metrics
  SSE      <- sum((predicted - true)^2)
  SST      <- sum((true - mean(true))^2)
  R_square <- 1 - SSE / SST
  RMSE     <- sqrt(SSE/nrow(df))
  # Save metrics
  data.frame(RMSE = RMSE, Rsquare = R_square)
}

# Downsampling
control.logloss.ds <- trainControl(method          = "repeatedcv",
                                   number          = 5,
                                   repeats         = 5,
                                   summaryFunction = mnLogLoss,
                                   classProbs      = TRUE,
                                   sampling        = "down")

# Create grid of parameters
par_grid <-  expand.grid(alpha = 0, lambda = seq(3.5e-7, 1.5e-5, length.out = 6))

# Set seed
set.seed(538)
```

### Step 2: Format data

```{r, echo = `FALSE`}
# Format historical dataset
df.historical <- df.historical %>%
  as.data.frame(.) %>%
  mutate(year         = as.character(year),
         month        = as.character(month),
         season       = as.factor(season),
         summerkill   = as.numeric(summerkill),
         lat          = as.numeric(lat),
         long         = as.numeric(long),
         state        = as.factor(state),
         population   = as.numeric(population),
         mean_bot_z   = as.numeric(mean_bot_z),
         mean_surf_z  = as.numeric(mean_surf_z),
         mean_air_z   = as.numeric(mean_air_z),
         ice_duration = as.numeric(ice_duration)) %>%
  mutate_at(c(30:54, 58:104), ~replace_na(., 0))

# Format historical dataset
df.historical2 <- df.historical2 %>%
  as.data.frame(.) %>%
  mutate(year         = as.character(year),
         month        = as.character(month),
         season       = as.factor(season),
         summerkill   = as.factor(summerkill),
         lat          = as.numeric(lat),
         long         = as.numeric(long),
         state        = as.factor(state),
         population   = as.numeric(population),
         mean_bot_z   = as.numeric(mean_bot_z),
         mean_surf_z  = as.numeric(mean_surf_z),
         mean_air_z   = as.numeric(mean_air_z),
         ice_duration = as.numeric(ice_duration))

# Format future dataset
df.future <- df.future %>%
  as.data.frame(.) %>%
  mutate(year         = as.character(year),
         month        = as.character(month),
         season       = as.factor(season),
         lat          = as.numeric(lat),
         long         = as.numeric(long),
         state        = as.factor(state),
         population   = as.numeric(population),
         mean_bot_z   = as.numeric(mean_bot_z),
         mean_surf_z  = as.numeric(mean_surf_z),
         mean_air_z   = as.numeric(mean_air_z),
         ice_duration = as.numeric(ice_duration))

# Format snow dataset
df.snow <- df.snow %>%
  as.data.frame(.) %>%
  mutate(year  = as.character(year),
         month = as.character(month))

# Format all fishkill data
fishkill.all <- df.historical %>%
  dplyr::mutate(fishkill = ifelse(anthropogenic == 1 | infectious == 1 | summerkill == 1 | winterkill == 1 | unknown == 1, 1, 0)) %>%
  dplyr::mutate(fishkill_cause = ifelse(anthropogenic == 1, "Anthropogenic",
                                        ifelse(infectious == 1, "Infectious",
                                               ifelse(summerkill == 1, "Summerkill",
                                                      ifelse(winterkill == 1, "Winterkill",
                                                             ifelse(unknown == 1, "Unknown", "Summer Non-event")))))) %>%
  dplyr::mutate(fishkill       = as.factor(fishkill),
                fishkill_cause = as.factor(fishkill_cause))

# Format summerkill data
fishkill.summerkill <- df.historical %>%
  filter(month %in% c("6", "7", "8")) %>%
  dplyr::mutate(fishkill = ifelse(anthropogenic == 1 | infectious == 1 | summerkill == 1 | winterkill == 1 | unknown == 1, 1, 0)) %>%
  dplyr::mutate(fishkill_cause = ifelse(anthropogenic == 1, "Anthropogenic",
                                        ifelse(infectious == 1, "Infectious",
                                               ifelse(summerkill == 1, "Summerkill",
                                                      ifelse(winterkill == 1, "Winterkill",
                                                             ifelse(unknown == 1, "Unknown", "Summer Non-event")))))) %>%
  dplyr::mutate(fishkill       = as.factor(fishkill),
                fishkill_cause = as.factor(fishkill_cause))

# Format winterkill data
fishkill.winterkill <- df.historical %>%
  filter(month %in% c("12", "1", "2")) %>%
  dplyr::mutate(fishkill = ifelse(anthropogenic == 1 | infectious == 1 | summerkill == 1 | winterkill == 1 | unknown == 1, 1, 0)) %>%
  dplyr::mutate(fishkill_cause = ifelse(anthropogenic == 1, "Anthropogenic",
                                        ifelse(infectious == 1, "Infectious",
                                               ifelse(summerkill == 1, "Summerkill",
                                                      ifelse(winterkill == 1, "Winterkill",
                                                             ifelse(unknown == 1, "Unknown", "Winter Non-event")))))) %>%
  dplyr::mutate(fishkill       = as.factor(fishkill),
                fishkill_cause = as.factor(fishkill_cause))

# Format snow data
fishkill.snow <- fishkill.winterkill %>%   
  mutate(lat_round  = round(lat, 1),
         long_round = round(long, 1)) %>%
  mutate(year  = as.character(year)) %>%
  mutate(month = as.character(month)) %>%
  inner_join(df.snow, by = c('year', 'month', 'lat_round', 'long_round')) %>%
  mutate(winterkill = ifelse(!is.na(fishkill_cause) & fishkill_cause == 'Winterkill', 1, 0))

# Format ice duration data
fishkill.ice <- fishkill.all  %>%
  group_by(year, lat, long) %>%
  dplyr::summarize(ice_duration = mean(ice_duration),
                   winterkill   = max(winterkill))

# Create figure theme
pretty <- theme_classic(base_size = 12) +
  theme(#axis.title.y      = element_text(margin = margin(r = 20)),
        axis.title.x      = element_text(margin = margin(t = 20)), 
        axis.text         = element_text(color = "black"), 
        axis.ticks        = element_line(color = "black"),
        axis.ticks.length = unit(.25, "cm"), 
        legend.key.size   = unit(1.5,  'lines'),
        legend.position   = "right")
```

### Step 3: Counts

```{r, echo = `FALSE`}
# Events by state
df.state <- df.historical %>%
  group_by(state) %>%
  dplyr::summarize(anthropogenic = sum(anthropogenic, na.rm = TRUE),
                   infectious    = sum(infectious, na.rm = TRUE),
                   summerkill    = sum(as.numeric(summerkill), na.rm = TRUE),
                   winterkill    = sum(winterkill, na.rm = TRUE),
                   unknown       = sum(unknown, na.rm = TRUE),
                   fishkills     = sum(c(anthropogenic, infectious, summerkill, winterkill, unknown, na.rm = TRUE)))

# Fishkills by lake
df.lake <- df.historical %>%
  subset(., anthropogenic >= 1 | infectious >= 1 | summerkill >= 1 | winterkill >= 1 | unknown >= 1) %>%
  group_by(lat, long) %>%
  dplyr::summarize(anthropogenic   = sum(anthropogenic),
                   infectious      = sum(infectious),
                   summerkill      = sum(summerkill),
                   winterkill      = sum(winterkill),
                   unknown         = sum(unknown),
                   fishkills       = sum(c(anthropogenic, infectious, summerkill, winterkill, unknown)))

# Count of lakes with single or multiple fishkills, mean, SE
sum(df.lake$fishkills == 1)
sum(df.lake$fishkills > 1)
mean(df.lake$fishkills)
sd(df.lake$fishkills) / sqrt(length(df.lake$fishkills))
max(df.lake$fishkills)
min(df.lake$fishkills)

# Calculate time between events that occurred in same lake
time.events <- time.events %>%
  mutate(month_after = month_after / 12) %>%
  dplyr::summarize(time_after = year_after + month_after)

# Summary statistics
mean(time.events$time_after)
sd(time.events$time_after) / sqrt(length(time.events$time_after))
median(time.events$time_after)
hist(time.events$time_after)

# Calculate time between summerkills that occurred in same lake
time.summerkills <- time.summerkills %>%
  mutate(month_after = month_after / 12) %>%
  dplyr::summarize(time_after = year_after + month_after)

# Summary statistics
mean(time.summerkills$time_after)
sd(time.summerkills$time_after) / sqrt(length(time.summerkills$time_after))
median(time.summerkills$time_after)
hist(time.summerkills$time_after)

# Count of waterbodies
length(unique(levels(as.factor(df.historical$site_id))))
length(unique(levels(as.factor(df.historical$site_id))))

# Fishkill causes by thermal category
df.thermal <- df.historical %>%
  subset(., anthropogenic >= 1 | infectious >= 1 | summerkill >= 1 | winterkill >= 1 | unknown >= 1) %>%
  reshape2::melt(., id.vars = c("state", "anthropogenic", "infectious", "summerkill", "winterkill", "unknown"),
                 measure.vars = c("cool_temp", "cold_temp", "warm_temp", "unknown_temp")) %>%
  group_by(variable, anthropogenic, infectious, summerkill, winterkill, unknown) %>%
  dplyr::summarize(value = sum(value)) %>%
  mutate(anthropogenic = ifelse(anthropogenic >= 1, value, 0),
         infectious    = ifelse(infectious >= 1,    value, 0),
         summerkill    = ifelse(summerkill >= 1,    value, 0),
         winterkill    = ifelse(winterkill >= 1,    value, 0),
         unknown       = ifelse(unknown >= 1,       value, 0),
         value         = NULL) %>%
  dplyr::rename(thermal.category = variable) %>%
  reshape2::melt(., id.vars = c("thermal.category"),
                 measure.vars = c("anthropogenic", "infectious", "summerkill",
                                  "winterkill", "unknown")) %>%
  dplyr::rename(cause = variable,
                count = value) %>%
  group_by(thermal.category, cause) %>%
  dplyr::summarize(count = sum(count))

# Number of thermal classifications available
sum(df.thermal$count)

# Fishkill causes by family
df.family <- df.historical %>%
  subset(., anthropogenic >= 1 | infectious >= 1 | summerkill >= 1 | winterkill >= 1 | unknown >= 1) %>%
  reshape2::melt(., id.vars = c("anthropogenic", "infectious", "summerkill", "winterkill", "unknown"),
                 measure.vars = c("centrarchidae", "percidae", "esocidae", "ictaluridae", "clupeidae",
                                  "cyprinidae", "salmonidae", "catostomidae", "sciaenidae",
                                  "osmeridae", "lepisosteidae", "acipenseridae", "amiidae",
                                  "cottoidae", "gasterosteidae", "gobiidae")) %>%
  group_by(variable, anthropogenic, infectious, summerkill, winterkill, unknown) %>%
  dplyr::summarize(value = sum(value)) %>%
  mutate(anthropogenic = ifelse(anthropogenic >= 1, value, 0),
         infectious    = ifelse(infectious >= 1,    value, 0),
         summerkill    = ifelse(summerkill >= 1,    value, 0),
         winterkill    = ifelse(winterkill >= 1,    value, 0),
         unknown       = ifelse(unknown >= 1,       value, 0),
         value         = NULL) %>%
  dplyr::rename(family = variable) %>%
  reshape2::melt(., id.vars = c("family"),
                 measure.vars = c("anthropogenic", "infectious", "summerkill",
                                  "winterkill", "unknown")) %>%
  dplyr::rename(cause = variable,
                count = value) %>%
  group_by(family, cause) %>%
  dplyr::summarize(count = as.numeric(sum(count))) %>%
  mutate(family = as.character(family),
         cause  = as.character(cause)) %>%
  .[with(., order(family, cause)), ]

# Summerkills by family
df.family %>%
  filter(cause == "summerkill") %>%
  dplyr::summarize(count = sum(count))
  
# Fishkills by family
df.family %>%
  group_by(family) %>%
  dplyr::summarize(count = sum(count))

# Fishkills with taxonomic information
df.historical %>%
  filter(anthropogenic >= 1 | infectious >= 1 |
           summerkill >= 1 | winterkill >= 1 |
           unknown >= 1, fish_families >= 1) %>%
  nrow(.)

# Fishkills with thermal category
df.historical %>%
  filter(warm_temp >= 1 | cool_temp >= 1 | cold_temp >= 1 | unknown_temp >= 1) %>%
  dplyr::summarize(warm_temp    = sum(warm_temp, na.rm = TRUE),
                   cool_temp    = sum(cool_temp, na.rm = TRUE),
                   cold_temp    = sum(cold_temp, na.rm = TRUE),
                   unknown_temp = sum(unknown_temp, na.rm = TRUE))

# Columaris
table(df.historical$cause.detail)
```

### Step 4: Means

```{r, echo = `FALSE`}
#################################################################
### Summer

# Summer non-events
summer.temp <- fishkill.summerkill %>%
  filter(summerkill == 0)

# Anthropogenic
anthro.temp <- fishkill.summerkill %>%
  filter(anthropogenic >= 1)

# Infectious
infect.temp <- fishkill.summerkill %>%
  filter(infectious >= 1)

# Summerkill
summerkill.temp <- fishkill.summerkill %>%
  filter(summerkill >= 1)

#################################################################
### Winter

# Winter non-event
winter.temp <- fishkill.winterkill %>%
  filter(winterkill == 0)

# Winterkill
winterkill.temp <- fishkill.winterkill %>%
  filter(winterkill >= 1)

#################################################################
### Anthropogenic

# Mean (% change)
((mean(anthro.temp$mean_air)  - mean(summer.temp$mean_air))  /  (mean(summer.temp$mean_air)) * 100)
((mean(anthro.temp$mean_surf) - mean(summer.temp$mean_surf)) / (mean(summer.temp$mean_surf)) * 100)
((mean(anthro.temp$mean_bot)  - mean(summer.temp$mean_bot))  / (mean(summer.temp$mean_bot))  * 100)

# Median (% change)
((median(anthro.temp$mean_air)  - median(summer.temp$mean_air))  /  (median(summer.temp$mean_air)) * 100)
((median(anthro.temp$mean_surf) - median(summer.temp$mean_surf)) / (median(summer.temp$mean_surf)) * 100)
((median(anthro.temp$mean_bot)  - median(summer.temp$mean_bot))  / (median(summer.temp$mean_bot))  * 100)

#################################################################
### Infectious

# Mean (% change)
((mean(infect.temp$mean_air)  - mean(summer.temp$mean_air))  /  (mean(summer.temp$mean_air)) * 100)
((mean(infect.temp$mean_surf) - mean(summer.temp$mean_surf)) / (mean(summer.temp$mean_surf)) * 100)
((mean(infect.temp$mean_bot)  - mean(summer.temp$mean_bot))  / (mean(summer.temp$mean_bot))  * 100)

# Median (% change)
((median(infect.temp$mean_air)  - median(summer.temp$mean_air))  /  (median(summer.temp$mean_air)) * 100)
((median(infect.temp$mean_surf) - median(summer.temp$mean_surf)) / (median(summer.temp$mean_surf)) * 100)
((median(infect.temp$mean_bot)  - median(summer.temp$mean_bot))  / (median(summer.temp$mean_bot))  * 100)

#################################################################
### Summerkill

# Mean (% change)
((mean(summerkill.temp$mean_air)  - mean(summer.temp$mean_air))  /  (mean(summer.temp$mean_air)) * 100)
((mean(summerkill.temp$mean_surf) - mean(summer.temp$mean_surf)) / (mean(summer.temp$mean_surf)) * 100)
((mean(summerkill.temp$mean_bot)  - mean(summer.temp$mean_bot))  / (mean(summer.temp$mean_bot))  * 100)

# Median (% change)
((median(summerkill.temp$mean_air)  - median(summer.temp$mean_air))  /  (median(summer.temp$mean_air)) * 100)
((median(summerkill.temp$mean_surf) - median(summer.temp$mean_surf)) / (median(summer.temp$mean_surf)) * 100)
((median(summerkill.temp$mean_bot)  - median(summer.temp$mean_bot))  / (median(summer.temp$mean_bot))  * 100)

#################################################################
### Winterkill

# Mean (% change)
((mean(winterkill.temp$mean_air)  - mean(winter.temp$mean_air))  /  (mean(winter.temp$mean_air)) * 100)
((mean(winterkill.temp$mean_surf) - mean(winter.temp$mean_surf)) / (mean(winter.temp$mean_surf)) * 100)
((mean(winterkill.temp$mean_bot)  - mean(winter.temp$mean_bot))  / (mean(winter.temp$mean_bot))  * 100)

# Median (% change)
((median(winterkill.temp$mean_air)  - median(winter.temp$mean_air))  /  (median(winter.temp$mean_air)) * 100)
((median(winterkill.temp$mean_surf) - median(winter.temp$mean_surf)) / (median(winter.temp$mean_surf)) * 100)
((median(winterkill.temp$mean_bot)  - median(winter.temp$mean_bot))  / (median(winter.temp$mean_bot))  * 100)

#################################################################

# Remove objects
rm(summer.temp, anthro.temp, infect.temp, summerkill.temp,
   winter.temp, winterkill.temp)
```

### Step 4: Mean temperatures

Compare mean temperatures of events and non-events

Use Welch's t-test if Levene's test (p < 0.05), then Dunnett test
Use ANOVA if Levene's test (p > 0.05), then Tukey HSD

```{r, echo = `FALSE`}
#################################################################
### Summer

# Summer non-event
summer.lakes <- fishkill.summerkill %>% 
  filter(fishkill_cause == 'Summer Non-event')
qqnorm(summer.lakes$mean_air)
qqnorm(summer.lakes$mean_surf)
qqnorm(summer.lakes$mean_bot)

# Anthropogenic
anthro.lakes <- fishkill.summerkill %>%
  filter(fishkill_cause == 'Anthropogenic')
qqnorm(anthro.lakes$mean_air)
qqnorm(anthro.lakes$mean_surf)
qqnorm(anthro.lakes$mean_bot)

# Summerkill
summerkill.lakes <- fishkill.summerkill %>%
  filter(fishkill_cause == 'Summerkill')
qqnorm(summerkill.lakes$mean_air)
qqnorm(summerkill.lakes$mean_surf)
qqnorm(summerkill.lakes$mean_bot)

# Infectious
infectious.lakes <- fishkill.summerkill %>%
  filter(fishkill_cause == 'Infectious')
qqnorm(infectious.lakes$mean_air)
qqnorm(infectious.lakes$mean_surf)
qqnorm(infectious.lakes$mean_bot)

#################################################################
### Winter

# Winter non-event
winter.lakes <- fishkill.winterkill %>%
  filter(fishkill_cause == 'Winter Non-event')
qqnorm(winter.lakes$mean_air)
qqnorm(winter.lakes$mean_surf)
qqnorm(winter.lakes$mean_bot)

# Winterkill
winterkill.lakes <- fishkill.winterkill %>%
  filter(fishkill_cause == 'Winterkill')
qqnorm(winterkill.lakes$mean_air)
qqnorm(winterkill.lakes$mean_surf)
qqnorm(winterkill.lakes$mean_bot)

# Remove objects
rm(summer.lakes, anthro.lakes, infectious.lakes, summerkill.lakes,
   winter.lakes, winterkill.lakes)
```

```{r, echo = `FALSE`}
#################################################################
### Air

# Levene test
leveneTest(mean_air ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(mean_air ~ fishkill_cause, fishkill.winterkill) # ANOVA

# Dunnett's test
DunnettTest(mean_air ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_air ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(mean_air ~ infectious, fishkill.summerkill)

# Summerkill
t.test(mean_air ~ summerkill, fishkill.summerkill)

# Winterkill
#winter.a <- aov(mean_air ~ fishkill_cause, fishkill.winterkill)
winter.a <- aov(mean_air ~ as.factor(winterkill), fishkill.winterkill)
summary(winter.a)
TukeyHSD(winter.a)

t.test(mean_air ~ winterkill, fishkill.winterkill)

#################################################################
### Surface water

# Levene test
leveneTest(mean_surf ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(mean_surf ~ fishkill_cause, fishkill.winterkill) # ANOVA (barely)

# Dunnett's test
DunnettTest(mean_surf ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_surf ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(mean_surf ~ infectious, fishkill.summerkill)

# Summerkill
t.test(mean_surf ~ summerkill, fishkill.summerkill)

# Winterkill
#winter.s <- aov(mean_surf ~ fishkill_cause, fishkill.winterkill)
winter.s <- aov(mean_surf ~ as.factor(winterkill), fishkill.winterkill)
summary(winter.s)
TukeyHSD(winter.s)

#################################################################
### Deep water

# Levene test
leveneTest(mean_bot ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(mean_bot ~ fishkill_cause, fishkill.winterkill) # ANOVA (barely)

# Dunnett's test
DunnettTest(mean_bot ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_bot ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(mean_bot ~ infectious, fishkill.summerkill)

# Summerkill
t.test(mean_bot ~ summerkill, fishkill.summerkill)

# Winterkill
#winter.b <- aov(mean_bot ~ fishkill_cause, fishkill.winterkill)
winter.b <- aov(mean_bot ~ as.factor(winterkill), fishkill.winterkill)
summary(winter.b)
TukeyHSD(winter.b)
```

### Step 4: z-scores

Compare z-scores for events and non-events, and check if event z-scores are non-zero

```{r, echo = `FALSE`}
#################################################################
### Summer

# Summer non-event
summer.lakes <- fishkill.summerkill %>% 
  filter(fishkill_cause == 'Summer Non-event')
qqnorm(summer.lakes$mean_air_z)
qqnorm(summer.lakes$mean_surf_z)
qqnorm(summer.lakes$mean_bot_z)

# Anthropogenic
anthro.lakes <- fishkill.summerkill %>%
  filter(fishkill_cause == 'Anthropogenic')
qqnorm(anthro.lakes$mean_air_z)
qqnorm(anthro.lakes$mean_surf_z)
qqnorm(anthro.lakes$mean_bot_z)

# Infectious
infectious.lakes <- fishkill.summerkill %>%
  filter(fishkill_cause == 'Infectious')
qqnorm(infectious.lakes$mean_air_z)
qqnorm(infectious.lakes$mean_surf_z)
qqnorm(infectious.lakes$mean_bot_z)

# Summerkill
summerkill.lakes <- fishkill.summerkill %>%
  filter(fishkill_cause == 'Summerkill')
qqnorm(summerkill.lakes$mean_air_z)
qqnorm(summerkill.lakes$mean_surf_z)
qqnorm(summerkill.lakes$mean_bot_z)

#################################################################
### Winter

# Winter non-event
winter.lakes <- fishkill.winterkill %>%
  filter(fishkill_cause == 'Winter Non-event')
qqnorm(winter.lakes$mean_air_z)
qqnorm(winter.lakes$mean_surf_z)
qqnorm(winter.lakes$mean_bot_z)

# Winterkill
winterkill.lakes <- fishkill.winterkill %>%
  filter(fishkill_cause == 'Winterkill')
qqnorm(winterkill.lakes$mean_air_z)
qqnorm(winterkill.lakes$mean_surf_z)
qqnorm(winterkill.lakes$mean_bot_z)

# Remove objects
rm(summer.lakes, anthro.lakes, infectious.lakes, summerkill.lakes,
   winter.lakes, winterkill.lakes)
```

```{r, echo = `FALSE`}
#################################################################
### Air

# Levene test
leveneTest(mean_air_z ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(mean_air_z ~ fishkill_cause, fishkill.winterkill) # ANOVA

# Dunnett's test
DunnettTest(mean_air_z ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_air_z ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(mean_air_z ~ infectious, fishkill.summerkill)

# Summerkill
t.test(mean_air_z ~ summerkill, fishkill.summerkill)

# Winterkill
winter.a <- aov(mean_air_z ~ fishkill_cause, fishkill.winterkill)
winter.a <- aov(mean_air ~ winterkill, fishkill.winterkill)
summary(winter.a)
TukeyHSD(winter.a)

#################################################################
### Surface water

# Levene test
leveneTest(mean_surf_z ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(mean_surf_z ~ fishkill_cause, fishkill.winterkill) # ANOVA (barely)

# Dunnett's test
DunnettTest(mean_surf_z ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_surf_z ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(mean_surf_z ~ infectious, fishkill.summerkill)

# Summerkill
t.test(mean_surf_z ~ summerkill, fishkill.summerkill)

# Winterkill
winter.s <- aov(mean_surf_z ~ fishkill_cause, fishkill.winterkill)
summary(winter.s)
TukeyHSD(winter.s)

#################################################################
### Deep water

# Levene test
leveneTest(mean_bot_z ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(mean_bot_z ~ fishkill_cause, fishkill.winterkill) # ANOVA (barely)

# Dunnett's test
DunnettTest(mean_bot_z ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_bot_z ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(mean_bot_z ~ infectious, fishkill.summerkill)

# Summerkill
t.test(mean_bot_z ~ summerkill, fishkill.summerkill)

# Winterkill
winter.b <- aov(mean_bot_z ~ fishkill_cause, fishkill.winterkill)
summary(winter.b)
TukeyHSD(winter.b)
```

```{r, echo = `FALSE`}
# Anthropogenic
ttest.a.z <- filter(fishkill.all, fishkill_cause == 'Anthropogenic')
t.test(ttest.a.z$mean_air_z,  mu = 0)
t.test(ttest.a.z$mean_surf_z, mu = 0)
t.test(ttest.a.z$mean_bot_z,  mu = 0)

# Infectious
ttest.i.z <- filter(fishkill.all, fishkill_cause == 'Infectious')
t.test(ttest.i.z$mean_air_z,  mu = 0)
t.test(ttest.i.z$mean_surf_z, mu = 0)
t.test(ttest.i.z$mean_bot_z,  mu = 0)

# Summerkill
ttest.s.z <- filter(fishkill.all, fishkill_cause == 'Summerkill')
t.test(ttest.s.z$mean_air_z,  mu = 0)
t.test(ttest.s.z$mean_surf_z, mu = 0)
t.test(ttest.s.z$mean_bot_z,  mu = 0)

# Winterkill
ttest.w.z <- filter(fishkill.all, fishkill_cause == 'Winterkill')
t.test(ttest.w.z$mean_air_z,  mu = 0)
t.test(ttest.w.z$mean_surf_z, mu = 0)
t.test(ttest.w.z$mean_bot_z,  mu = 0)
```

### Step 5: Snow and ice duration

Compare snow accumulation and ice duration for winterkills and winter non-events

```{r, echo = `FALSE`}
# Winter non-event
winter.lakes <- fishkill.winterkill %>% 
  filter(fishkill_cause == 'Winter Non-event')
qqnorm(winter.lakes$mean_air)
qqnorm(winter.lakes$mean_surf)
qqnorm(winter.lakes$mean_bot)

# Winterkill
winterkill.lakes <- fishkill.winterkill %>%
  filter(fishkill_cause == 'Winterkill')
qqplot(x = winter.lakes$mean_air_z,  y = winterkill.lakes$mean_air_z)
qqplot(x = winter.lakes$mean_surf_z, y = winterkill.lakes$mean_surf_z)
qqplot(x = winter.lakes$mean_bot_z,  y = winterkill.lakes$mean_bot_z)

# Ice duration
nonevent.ice   <- fishkill.ice %>% filter(winterkill == 0)
winterkill.ice <- fishkill.ice %>% filter(winterkill == 1)
qqplot(nonevent.ice$ice_duration, winterkill.ice$ice_duration)

# Remove objects
rm(winter.lakes, winterkill.lakes,
   nonevent.ice, winterkill.ice)
```

```{r, echo = `FALSE`}
# Levene test (snow)
leveneTest(snow ~ as.factor(winterkill), fishkill.snow) # ANOVA / Tuk

# ANOVA (snow)
anova.w.snow <- aov(snow ~ as.factor(winterkill), fishkill.snow)
summary(anova.w.snow)
TukeyHSD(anova.w.snow)

# Levene test (ice duration)
leveneTest(ice_duration ~ as.factor(winterkill), fishkill.ice) # ANOVA / Tuk (barely)

# ANOVA (ice duration)
anova.w.ice <- aov(ice_duration ~ as.factor(winterkill), fishkill.ice)
summary(anova.w.ice)
TukeyHSD(anova.w.ice)

# Remove objects
rm(anova.w.snow, anova.w.ice)
```

### Step 4: Lake size and secchi depth

```{r, echo = `FALSE`}
#################################################################
### Lake size (km^2)

# Levene test
leveneTest(size_km2 ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(size_km2 ~ fishkill_cause, fishkill.winterkill) # ANOVA

# Dunnett's test
DunnettTest(size_km2 ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(mean_air ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(size_km2 ~ infectious, fishkill.summerkill)

# Summerkill
t.test(size_km2 ~ summerkill, fishkill.summerkill)

# Winterkill
#winter.a <- aov(size_km2 ~ fishkill_cause, fishkill.winterkill)
winter.a <- aov(size_km2 ~ as.factor(winterkill), fishkill.winterkill)
summary(winter.a)
TukeyHSD(winter.a)

t.test(size_km2 ~ winterkill, fishkill.winterkill)

#################################################################
### Secchi depth

# Levene test
leveneTest(secchi_m ~ fishkill_cause, fishkill.summerkill) # Welch
leveneTest(secchi_m ~ fishkill_cause, fishkill.winterkill) # ANOVA (barely)

# Dunnett's test
DunnettTest(secchi_m ~ fishkill_cause, fishkill.summerkill, control = 'Summer Non-event')

# Anthropogenic (not enough observations)
#t.test(secchi_m ~ anthropogenic, fishkill.summerkill)

# Infectious
t.test(secchi_m ~ infectious, fishkill.summerkill)

# Summerkill
t.test(secchi_m ~ summerkill, fishkill.summerkill)

# Winterkill
#winter.s <- aov(secchi_m ~ fishkill_cause, fishkill.winterkill)
winter.s <- aov(secchi_m ~ as.factor(winterkill), fishkill.winterkill)
summary(winter.s)
TukeyHSD(winter.s)
```

### Step 6: Thermal categories

Linear regression with CTmax or thermal preference and difference with maximum surface water temp of summerkills

```{r, echo = `FALSE`}
# Cold-water fish
cold.temp <- table(fishkill.all$fishkill_cause, fishkill.all$cold_temp)
cold.temp <- cold.temp[,2]

# Cool-water fish
cool.temp <- table(fishkill.all$fishkill_cause, fishkill.all$cool_temp)
cool.temp <- cool.temp[,2]

# Warm-water fish
warm.temp <- table(fishkill.all$fishkill_cause, fishkill.all$warm_temp)
warm.temp <- warm.temp[,2]

# Combine tables
temp.table <- cold.temp %>%
  cbind(., cool.temp, warm.temp) %>%
  .[-3,]

# Fix columns names
colnames(temp.table) <- c("cold.temp", "cool.temp", "warm.temp")

# Fisher exact test
#fisher.test(temp.table)
fisher_test(temp.table, detailed = TRUE)

# Causes
anthropogenic <- temp.table[c("Anthropogenic"),]
infectious    <- temp.table[c("Infectious"),]
summerkill    <- temp.table[c("Summerkill"),]
winterkill    <- temp.table[c("Winterkill"),]

# Thermal categories
total <- c(sum(temp.table[,c("cold.temp")]),
           sum(temp.table[,c("cool.temp")]),
           sum(temp.table[,c("warm.temp")]))

# Non-causes
non.anthropogenic <- total - anthropogenic
non.infectious    <- total - infectious
non.summerkill    <- total - summerkill
non.winterkill    <- total - winterkill

# Anthropogenic
table.a <- as.table(rbind(anthropogenic, non.anthropogenic))
dimnames(table.a) <- list(Anthropogenic = c("Yes", "No"),
                         Category   = c("Cold", "Cool", "Warm"))

fisher_test(table.a, detailed = TRUE)
pairwise_fisher_test(table.a, p.adjust.method = "bonferroni")

# Infectious
table.i <- as.table(rbind(infectious, non.infectious))
dimnames(table.i) <- list(Infectious = c("Yes", "No"),
                          Category   = c("Cold", "Cool", "Warm"))

fisher_test(table.i, detailed = TRUE)
pairwise_fisher_test(table.i, p.adjust.method = "bonferroni")

# Summerkill
table.s <- as.table(rbind(summerkill, non.summerkill))
dimnames(table.s) <- list(Summerkill = c("Yes", "No"),
                          Category   = c("Cold", "Cool", "Warm"))

fisher_test(table.s, detailed = TRUE)
pairwise_fisher_test(table.s, p.adjust.method = "bonferroni")

# Winterkill
table.w <- as.table(rbind(winterkill, non.winterkill))
dimnames(table.w) <- list(Winterkill = c("Yes", "No"),
                         Category   = c("Cold", "Cool", "Warm"))

fisher_test(table.w, detailed = TRUE)
pairwise_fisher_test(table.w)

# Multiple comparisons
#fisher.multcomp(temp.table, p.method = "bonferroni")

# Reformat CTmax data
thermal <- thermal %>%
  rename_all(tolower) %>%
  mutate_all(tolower) %>%
  mutate(pref_min = as.numeric(pref_min),
         pref_max = as.numeric(pref_max),
         crit_min = as.numeric(crit_min),
         crit_max = as.numeric(crit_max)) %>%
  dplyr::select(., c("family", "pref_min", "pref_max", "crit_min", "crit_max",)) %>%
  na.omit(.) %>%
  group_by(family) %>%
  dplyr::summarize(pref_min = mean(pref_min),
                   pref_max = mean(pref_max),
                   crit_min = mean(crit_min),
                   crit_max = mean(crit_max))

# Centrarchidae
thermal.centrarchidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(centrarchidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%
  mutate(family = as.character("centrarchidae"),
         pref_min = thermal$pref_min[which(thermal$family == "centrarchidae")],
         pref_max = thermal$pref_max[which(thermal$family == "centrarchidae")],
         crit_min = thermal$crit_min[which(thermal$family == "centrarchidae")],
         crit_max = thermal$crit_max[which(thermal$family == "centrarchidae")])

# Percidae
thermal.percidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(percidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("percidae"),
         pref_min = thermal$pref_min[which(thermal$family == "percidae")],
         pref_max = thermal$pref_max[which(thermal$family == "percidae")],
         crit_min = thermal$crit_min[which(thermal$family == "percidae")],
         crit_max = thermal$crit_max[which(thermal$family == "percidae")])

# Esocidae
thermal.esocidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(esocidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("esocidae"),
         pref_min = thermal$pref_min[which(thermal$family == "esocidae")],
         pref_max = thermal$pref_max[which(thermal$family == "esocidae")],
         crit_min = thermal$crit_min[which(thermal$family == "esocidae")],
         crit_max = thermal$crit_max[which(thermal$family == "esocidae")])

# Ictaluridae
thermal.ictaluridae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(ictaluridae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("ictaluridae"),
         pref_min = thermal$pref_min[which(thermal$family == "ictaluridae")],
         pref_max = thermal$pref_max[which(thermal$family == "ictaluridae")],
         crit_min = thermal$crit_min[which(thermal$family == "ictaluridae")],
         crit_max = thermal$crit_max[which(thermal$family == "ictaluridae")])

# Salmonidae
thermal.salmonidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(salmonidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("salmonidae"),
         pref_min = thermal$pref_min[which(thermal$family == "salmonidae")],
         pref_max = thermal$pref_max[which(thermal$family == "salmonidae")],
         crit_min = thermal$crit_min[which(thermal$family == "salmonidae")],
         crit_max = thermal$crit_max[which(thermal$family == "salmonidae")])

# Catostomidae
thermal.catostomidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(catostomidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("catostomidae"),
         pref_min = thermal$pref_min[which(thermal$family == "catostomidae")],
         pref_max = thermal$pref_max[which(thermal$family == "catostomidae")],
         crit_min = thermal$crit_min[which(thermal$family == "catostomidae")],
         crit_max = thermal$crit_max[which(thermal$family == "catostomidae")])

# Sciaenidae
thermal.sciaenidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(sciaenidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("sciaenidae"),
         pref_min = thermal$pref_min[which(thermal$family == "sciaenidae")],
         pref_max = thermal$pref_max[which(thermal$family == "sciaenidae")],
         crit_min = thermal$crit_min[which(thermal$family == "sciaenidae")],
         crit_max = thermal$crit_max[which(thermal$family == "sciaenidae")])

# Clupeidae
thermal.clupeidae <- df.historical %>%
  filter(summerkill >= 1) %>%
  filter(clupeidae >= 1) %>%
  dplyr::select(., c("max_surf", "max_surf_z", "mean_surf",
                     "max_bot", "max_bot_z", "mean_bot",
                     "max_air", "max_air_z", "mean_air")) %>%  mutate(family = as.character("clupeidae"),
         pref_min = thermal$pref_min[which(thermal$family == "clupeidae")],
         pref_max = thermal$pref_max[which(thermal$family == "clupeidae")],
         crit_min = thermal$crit_min[which(thermal$family == "clupeidae")],
         crit_max = thermal$crit_max[which(thermal$family == "clupeidae")])

# Join CTmax df
thermal.families <- thermal.centrarchidae %>%
  rbind(., thermal.percidae) %>%
  rbind(., thermal.esocidae) %>%
  rbind(., thermal.ictaluridae) %>%
  rbind(., thermal.clupeidae) %>%
  rbind(., thermal.salmonidae) %>%
  rbind(., thermal.sciaenidae) %>%
  rbind(., thermal.catostomidae) %>%
  # Difference
  mutate(
    diff_crit_air    = max_air -  crit_min,
    diff_crit_surf   = max_surf - crit_min,
    diff_crit_bot    = max_bot -  crit_min,
    diff_pref_air    = max_air -  pref_min,
    diff_pref_surf   = max_surf - pref_min,
    diff_pref_bot    = max_bot -  pref_min) %>%
  # Pref temperature
  rowwise() %>%
  mutate(
    pref_mean = sum (pref_min, pref_max, na.rm = TRUE) / 2)

# Summarize family data
thermal.sum <- thermal.families %>%
  group_by(family) %>%
  dplyr::summarize(
    # Lyons et al. 2019
    pref_min         = pref_min,
    pref_max         = pref_max,
    crit_min         = crit_min,
    crit_max         = crit_max,
    # Max air
    max_air_n        = length(max_air),
    max_air_mean     = mean(max_air),
    max_air_sd       = sd(max_air),
    max_air_se       = max_air_sd / sqrt(max_air_n),
    # Max air (z-score)
    max_air_z_n      = length(max_air_z),
    max_air_z_mean   = mean(max_air_z),
    max_air_z_sd     = sd(max_air_z),
    max_air_z_se     = max_air_z_sd / sqrt(max_air_z_n),
    # Max surface water
    max_surf_n       = length(max_surf),
    max_surf_mean    = mean(max_surf),
    max_surf_sd      = sd(max_surf),
    max_surf_se      = max_surf_sd / sqrt(max_surf_n),
    # Max surface water (z-score)
    max_surf_z_n     = length(max_surf_z),
    max_surf_z_mean  = mean(max_surf_z),
    max_surf_z_sd    = sd(max_surf_z),
    max_surf_z_se    = max_surf_z_sd / sqrt(max_surf_z_n),
    # Max bottom water
    max_bot_n        = length(max_bot),
    max_bot_mean     = mean(max_bot),
    max_bot_sd       = sd(max_bot),
    max_bot_se       = max_bot_sd / sqrt(max_bot_n),
    # Max bottom water (z-score)
    max_bot_z_n      = length(max_bot_z),
    max_bot_z_mean   = mean(max_bot_z),
    max_bot_z_sd     = sd(max_bot_z),
    max_bot_z_se     = max_bot_z_sd / sqrt(max_bot_z_n)) %>%
  distinct(.)

# Linear regression CTmax ~ diff w/ max surf
summary(lm(crit_min ~ diff_crit_surf, thermal.families))

# Linear regression CTmax ~ max_surf_z
summary(lm(crit_min ~ max_surf_z, thermal.families))

# Remove objects
rm(df.family, df.lake, df.snow, df.thermal,
   fishkill.all, fishkill.ice, fishkill.snow,
   fishkill.summerkill, fishkill.winterkill,
   temp.table,
   anthropogenic, non.anthropogenic,
   infectious, non.infectious,
   summerkill, non.summerkill,
   winterkill, non.winterkill,
   cold.temp, cool.temp, warm.temp,
   table.a, table.i, table.s, table.w,
   total, thermal.catostomidae, thermal.centrarchidae,
   thermal.clupeidae, thermal.esocidae, thermal.families,
   thermal.ictaluridae, thermal.percidae, thermal.salmonidae,
   thermal.sciaenidae, thermal.sum, thermal, time.events)
```

### Step 8: Lake size and productivity
```{r, echo = FALSE}
#################################################################
### Lake size

# Format data
summerkill.size <- summerkill.temp %>%
  dplyr::select(., c(site_id, size_km2)) %>%
  distinct(.)

# Format data
summer.size <- summer.temp %>%
  dplyr::select(., c(site_id, size_km2)) %>%
  distinct(.)

### Lake size

# Mean (% change)
((mean(summerkill.size$size_km2)  - mean(summer.size$size_km2))  /  (mean(summer.size$size_km2)) * 100)

# Median (% change)
((median(summerkill.size$size_km2)  - median(summer.size$size_km2))  /  (median(summer.size$size_km2)) * 100)

#################################################################
### Productivity

# Format data
summerkill.prod <- summerkill.temp %>%
  dplyr::select(., c(site_id, secchi_m)) %>%
  distinct(.)

# Format size data
summer.prod <- summer.temp %>%
  dplyr::select(., c(site_id, secchi_m)) %>%
  distinct(.)

# Mean (% change)
((mean(summerkill.prod$secchi_m)  - mean(summer.prod$secchi_m))  /  (mean(summer.prod$secchi_m)) * 100)

# Median (% change)
((median(summerkill.prod$secchi_m)  - median(summer.prod$secchi_m))  / (median(summer.prod$secchi_m))  * 100)

```

### Step 7: Historical probabilities

```{r, echo = `FALSE`}
# Subset variables 
ridge_a2_historical <- df.historical2 %>%
  dplyr::select(., c(lat, long, season, population,
                     max_air, mean_air, min_air,
                     max_air_z, mean_air_z, min_air_z,
                     air_quad_temp, ice_duration, precip,
                     secchi_m, size_km2))

# Replace summerkills with 0s
ridge_a2_historical$summerkill <- 0

# Create dummy model matrix
newx.a2 <- model.matrix(summerkill ~ ., ridge_a2_historical)

# Calculate probabilities
ridge_a2_historical$prob <- as.vector(predict(ridge_a2,
                                              newx = newx.a2, type = "prob",
                                              s = ridge_a2$bestTune$lambda))

# Reformat probabilities
ridge_a2_historical <- ridge_a2_historical %>%
  mutate(prob_pos = prob$pos,
         prob     = NULL,
         lat      = df.historical$lat,
         long     = df.historical$long) %>%
  dplyr::rename(prob = prob_pos)

## Quick plot
#ggplot() +
#  geom_point(ridge_a2_historical,
#             mapping = aes(x = air_pc1, y = prob),
#             alpha = 0.002)

# Save probabilities
write.csv(ridge_a2_historical, "data/processed/ridge_a2_historical.csv")
```

```{r, echo = `FALSE`}
# Subset variables 
ridge_w1_historical <- df.historical2 %>%
  dplyr::select(., c(lat, long, season, population,
                     mean_surf, mean_surf_z,
                     ice_duration, precip,
                     secchi_m, size_km2))

# Replace summerkills with 0s
ridge_w1_historical$summerkill <- 0

# Create dummy model matrix
newx.w1 <- model.matrix(summerkill ~ ., ridge_w1_historical)

# Calculate probabilities
ridge_w1_historical$prob <- as.vector(predict(ridge_w1,
                                              newx = newx.w1, type = "prob",
                                              s = ridge_w1$bestTune$lambda))

# Reformat probabilities
ridge_w1_historical <- ridge_w1_historical %>%
  mutate(prob_pos = prob$pos,
         prob     = NULL,
         lat      = df.historical$lat,
         long     = df.historical$long) %>%
  dplyr::rename(prob = prob_pos)

## Quick plot
#ggplot() +
#  geom_point(ridge_w1_historical,
#             mapping = aes(x = mean_surf, y = prob),
#             alpha = 0.002)

# Save probabilities
write.csv(ridge_w1_historical, "data/processed/ridge_w1_historical.csv")
```

### Step 9: Future probabilities

```{r, echo = `FALSE`}
# Future a2 model
a2.future <- summerkill ~ lat + long + season + population +
  max_air + mean_air + min_air + max_air_z + mean_air_z + min_air_z +
  air_quad_temp + precip + secchi_m + size_km2

# Future a2 model
future_ridge_a2_ds <- train(a2.future, 
                            data      = df.historical2, 
                            method    = "glmnet",
                            metric    = "logLoss",
                            tuneGrid  = par_grid,
                            trControl = control.logloss.ds)

# Subset variables 
ridge_a2_future <- df.future %>%
  dplyr::select(., c(lat, long, season, population,
                     max_air, mean_air, min_air,
                     max_air_z, mean_air_z, min_air_z,
                     air_quad_temp, precip, secchi_m, size_km2))

# Calculate probabilities
ridge_a2_future$prob <- as.vector(predict(future_ridge_a2_ds,
                                          newdata = ridge_a2_future, type = "prob",
                                          s = future_ridge_a2_ds$bestTune$lambda))

# Reformat probabilities
ridge_a2_future <- ridge_a2_future %>%
  mutate(prob_pos = prob$pos,
         prob     = NULL,
         lat      = df.future$lat,
         long     = df.future$long) %>%
  dplyr::rename(prob = prob_pos)

# Save probabilities
write.csv(ridge_a2_future, "data/processed/ridge_a2_future.csv")
```

```{r, echo = `FALSE`}
# Future w1 model
w1.future <- summerkill ~ lat + long + season + population +
  mean_surf + mean_surf_z + precip + secchi_m + size_km2

# Future w1 model
future_ridge_w1_ds <- train(w1.future, 
                            data      = df.historical2, 
                            method    = "glmnet",
                            metric    = "logLoss",
                            tuneGrid  = par_grid,
                            trControl = control.logloss.ds)

# Subset variables 
ridge_w1_future <- df.future %>%
  dplyr::select(., lat, long, season, population,
                mean_surf, mean_surf_z, precip, secchi_m, size_km2)

# Calculate probabilities
ridge_w1_future$prob <- as.vector(predict(future_ridge_w1_ds,
                                          newdata = ridge_w1_future, type = "prob",
                                          s = future_ridge_w1_ds$bestTune$lambda))

# Reformat probabilities
ridge_w1_future <- ridge_w1_future %>%
  mutate(prob_pos = prob$pos,
         prob     = NULL,
         lat      = df.future$lat,
         long     = df.future$long) %>%
  dplyr::rename(prob = prob_pos)

# Save probabilities
write.csv(ridge_w1_future, "data/processed/ridge_w1_future.csv")
```

### Step 10: Moran's tests

```{r, echo = `FALSE`}
# Historical (2003-2013)
prob.a <- df.historical %>%
  dplyr::select(lat, long, year) %>%
  mutate(prob_air   = ridge_a2_historical$prob,
         prob_water = ridge_w1_historical$prob,
         no_air     = 1 - prob_air,
         no_water   = 1 - prob_water) %>%
  group_by(lat, long) %>%
  dplyr::summarize(prob_air   = (1 - prod(no_air)),
                   prob_water = (1 - prod(no_water)),
                   V1   = mean(long),
                   V2   = mean(lat)) 

# Future (2041-2059)
prob.b <- df.future %>%
  dplyr::select(lat, long, year) %>%
  mutate(prob_air   = ridge_a2_future$prob,
         prob_water = ridge_w1_future$prob,
         no_air     = 1 - prob_air,
         no_water   = 1 - prob_water) %>%
  filter(year > 2030 & year < 2070) %>%
  group_by(lat, long) %>%
  dplyr::summarize(prob_air   = (1 - prod(no_air)),
                   prob_water = (1 - prod(no_water)),
                   V1   = mean(long),
                   V2   = mean(lat)) 

# Future (2081-2099)
prob.c <- df.future %>%
  dplyr::select(lat, long, year) %>%
  mutate(prob_air   = ridge_a2_future$prob,
         prob_water = ridge_w1_future$prob,
         no_air     = 1 - prob_air,
         no_water   = 1 - prob_water) %>%
  filter(year > 2070) %>%
  group_by(lat, long) %>%
  dplyr::summarize(prob_air   = (1 - prod(no_air)),
                   prob_water = (1 - prod(no_water)),
                   V1   = mean(long),
                   V2   = mean(lat)) 

# Convert coordinates to SpatialPoints
coordinates(prob.a) <- c('V1', 'V2')
coordinates(prob.b) <- c('V1', 'V2')
coordinates(prob.c) <- c('V1', 'V2')

# Calculate weights
nb.a <- knn2nb(knearneigh(coordinates(coordinates(prob.a)), longlat = TRUE, k = 2))
nb.b <- knn2nb(knearneigh(coordinates(coordinates(prob.b)), longlat = TRUE, k = 2))
nb.c <- knn2nb(knearneigh(coordinates(coordinates(prob.c)), longlat = TRUE, k = 2))

# Add weights
weights.a <- nb2listw(nb.a)
weights.b <- nb2listw(nb.b)
weights.c <- nb2listw(nb.c)

# Moran's tests for air temperature
air.prob.a <- moran.test(prob.a$prob_air, weights.a)
air.prob.b <- moran.test(prob.b$prob_air, weights.b)
air.prob.c <- moran.test(prob.c$prob_air, weights.c)

# Print results
air.prob.a
air.prob.b
air.prob.c

# Moran's tests for water temperature
water.prob.a <- moran.test(prob.a$prob_water, weights.a)
water.prob.b <- moran.test(prob.b$prob_water, weights.b)
water.prob.c <- moran.test(prob.c$prob_water, weights.c)

# Print results
water.prob.a
water.prob.b
water.prob.c
```

######################################
Proceed to `07_figures.Rmd`
######################################
